+++ R/effect.R
@@ -82,8 +82,9 @@ add_mappers <- function(...) {
 #' \describe{
 #' \item{\eqn{c}{c}}{is the *intercept*}
 #' \item{\eqn{x}{x}}{is a *covariate*}
-#' \item{\eqn{\beta}{beta}}{is a *latent variable* associated with \eqn{x}{x} and}
-#' \item{\eqn{\psi = \beta x}{psi = beta * x}}{ is called the *effect* of \eqn{x}{x}}
+#' \item{\eqn{\beta}{beta}}{is a *latent variable* associated with \eqn{x}{x}
+#' and} \item{\eqn{\psi = \beta x}{psi = beta * x}}{ is called the *effect* of
+#' \eqn{x}{x}}
 #' }
 #'
 #' A problem that arises when using this kind of R formula is that it does not